+++ gulp.config.json
@@ -1,7 +1,7 @@
 {
   "paths": {
-    "src": "../src/",
-    "dist": "../dist/"
+    "src": "src/",
+    "dist": "dist/"
   },
   "styles": {
     "entryFiles": ["./main.scss"],
@@ -9,7 +9,7 @@
   },
   "scripts": {
     "entryFiles": ["./index.js"],
-    "watch": "src/scripts/**/*.js"
+    "watch": "src/js/**/*.js"
   },
   "env": "development",
   "vendorScripts": []
