+++ src/js/components/footer.js
@@ -1,0 +1,4 @@
+export function footer() {
+    console.log('FooterJS 2022');
+}
+
