+++ src/js/components/header.js
@@ -1,0 +1,4 @@
+export function header() {
+    console.log('HeaderJS')
+  }
+
