--- src/js/footerComponent.js
@@ -1,3 +1,0 @@
-export function footer() {
-  console.log('FooterJS')
-}
