--- src/js/headerComponent.js
@@ -1,3 +1,0 @@
-export function header() {
-  console.log('HeaderJS')
-}
